Refactor: Moved vector functions into pyxem.utils.vectors.py
